enhancement/Speakers-block-to-container
