--- src/App/blocks/Speakers/index.ts
@@ -1,2 +1,0 @@
-export { default } from './connected';
-export * from './Speakers';
