+++ src/App/containers/Speakers/Speakers.tsx
@@ -1,13 +1,128 @@
 import React, {
+	ContextType,
 	Component
 } from 'react';
-import Speakers from '~/blocks/Speakers';
+import {
+	withRouter
+} from 'react-router-dom';
+import {
+	I18nContext,
+	__x
+} from 'i18n-for-react';
+import {
+	getTalkTypes
+} from '~/services/i18n';
+import SpeakerModal from '~/blocks/SpeakerModal';
+import Section from '~/components/Section';
+import ToggleNav, {
+	ToggleNavLink
+} from '~/components/ToggleNav';
+import ProfileCard from '~/components/ProfileCard';
+import Badge from '~/components/Badge';
+import Loading from '~/components/Loading';
+import {
+	addSearchParams
+} from '~/blocks/common/router';
+import {
+	IProps
+} from './types';
+import {
+	style,
+	classes
+} from './Speakers.st.css';
 
-export default class SpeakersContainer extends Component {
+export class Speakers extends Component<IProps> {
+
+	static contextType = I18nContext;
+
+	context!: ContextType<typeof I18nContext>;
 
 	render() {
+
+		const {
+			className,
+			location: {
+				search
+			},
+			actionsReady,
+			selectSpeakersByType,
+			selectSpeaker
+		} = this.props;
+		const {
+			context
+		} = this;
+		const type = new URLSearchParams(search).get('type');
+		const nav = getTalkTypes(context);
+		const speakers = selectSpeakersByType(type);
+
 		return (
-			<Speakers/>
+			<Section
+				className={style(classes.root, className)}
+			>
+				<div
+					className={classes.group}
+				>
+					<h2>
+						{__x`speakers.title`}
+					</h2>
+				</div>
+				<ToggleNav
+					className={classes.nav}
+				>
+					{nav.map(item => (
+						<ToggleNavLink
+							key={item.type}
+							to={`/speakers?type=${item.type}`}
+						>
+							{item.label}
+						</ToggleNavLink>
+					))}
+				</ToggleNav>
+				<ul
+					className={classes.list}
+				>
+					{speakers.map(item => (
+						<li
+							key={item.id}
+						>
+							<ProfileCard
+								{...item}
+								badge={item.badge && (
+									<Badge
+										color='pink'
+									>
+										{item.badge}
+									</Badge>
+								)}
+								to={{
+									search: addSearchParams(search, {
+										id: item.id,
+										type
+									})
+								}}
+							/>
+						</li>
+					))}
+				</ul>
+				{actionsReady ? (
+					<SpeakerModal
+						getSpeaker={selectSpeaker}
+					/>
+				) : (
+					<Loading/>
+				)}
+			</Section>
 		);
 	}
+
+	componentDidMount() {
+
+		const {
+			fetchSpeakers
+		} = this.props;
+
+		fetchSpeakers();
+	}
 }
+
+export default withRouter(Speakers);
