+++ src/App/containers/Speakers/connected.ts
